idk lol

+++ app/Http/Controllers/SettingsController.php
@@ -11,8 +11,12 @@ class SettingsController extends Controller
     public function index(Request $request, User $user)
     {
         $updated_fields = []; 
+        // return response()->json(['message' =>$request->all()], 200);
+
         if ($request->hasFile('profile_picture')) {
-            Storage::disk('public')->delete($user->profile_picture);
+            if ($user->profile_picture) {
+                Storage::disk('public')->delete($user->profile_picture);
+            }
 
             $file = $request->file('profile_picture');
             $path = $file->store('profile_pictures', 'public');
